requirements.txt
cozepy>=0.20.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-multipart>=0.0.9
numpy>=1.26.0
pandas>=2.1.0
plotly>=5.22.0
openpyxl>=3.1.2
xlrd>=2.0.1
